dedup: share one stream-JSON parser between cos-runner and agentCliSpawning (#6827)
The Claude stream-JSON parser existed twice with identical behavior in
server/cos-runner/streamJsonParser.js and server/services/agentCliSpawning.js.
Every parser change had to land in both copies (three times so far), and a
fix landing in only one would silently drift agent transcripts and PR-body
summaries between the runner-spawned path and the in-process path.

Move the pure, import-free module to server/lib/streamJsonParser.js and have
both spawn paths import it. agentCliSpawning.js keeps re-exporting
createStreamJsonParser/summarizeToolInput/safeParse under the same names for
existing callers. Folded the two test files into one, keeping the more
thorough getFinalResult coverage from agentCliSpawning.test.js.

Closes #6827
