+++ server/services/agentCliSpawning.js
@@ -48,207 +48,13 @@ import { isHostShuttingDown, HOST_SHUTDOWN_REASON } from '../lib/hostShutdown.js
 import { ensureOllamaAgentContext } from './ollamaAgentContext.js';
 import { isOllamaBackedProvider } from './providers.js';
 import { isPublicReviewRestrictedProfile } from '../lib/agentExecutionProfiles.js';
+import { createStreamJsonParser, summarizeToolInput } from '../lib/streamJsonParser.js';
 
 const AGENTS_DIR = PATHS.cosAgents;
 
-/**
- * Summarize tool input into a concise description for display.
- * Extracts the most relevant parameter from each tool type.
- */
-export function summarizeToolInput(toolName, input) {
-  if (!input || typeof input !== 'object') return '';
-  const shorten = (p) => {
-    if (!p || typeof p !== 'string') return '';
-    const parts = p.split('/').filter(Boolean);
-    return parts.length > 2 ? `…/${parts.slice(-2).join('/')}` : p;
-  };
-  switch (toolName) {
-    case 'Read':
-      return shorten(input.file_path);
-    case 'Edit':
-      return shorten(input.file_path);
-    case 'Write':
-      return shorten(input.file_path);
-    case 'Glob':
-      return input.pattern || '';
-    case 'Grep':
-      return `"${(input.pattern || '').substring(0, 60)}"${input.path ? ` in ${shorten(input.path)}` : ''}`;
-    case 'Bash': {
-      const cmd = input.command || input.description || '';
-      return cmd.substring(0, 80);
-    }
-    case 'Task':
-      return input.description || '';
-    case 'WebFetch':
-      return shorten(input.url || '');
-    case 'WebSearch':
-      return `"${(input.query || '').substring(0, 60)}"`;
-    case 'TodoWrite':
-      return input.todos?.length ? `${input.todos.length} items` : '';
-    case 'NotebookEdit':
-      return shorten(input.notebook_path);
-    case 'Skill':
-      return input.skill || '';
-    default:
-      return '';
-  }
-}
-
+// Re-exported for existing callers/tests (server/lib/cursor.js, agentCliSpawning.test.js).
+export { createStreamJsonParser, summarizeToolInput };
 export const safeParse = (str) => safeJSONParse(str, null);
-
-/**
- * Create a Claude stream-json parser that extracts human-readable text from JSON stream events.
- * Returns a stateful parser with a `processChunk(data)` method that returns extracted text lines.
- * The parser handles:
- *   - content_block_delta: incremental text tokens as they stream
- *   - tool_use events: shows tool calls with input details (e.g. "🔧 Read …/services/api.js")
- *   - input_json_delta: accumulates tool input JSON for detailed summaries
- *   - content_block_stop: emits detailed tool summary when input is complete
- *   - result: final result text (used for output file)
- */
-export function createStreamJsonParser() {
-  let lineBuffer = '';
-  let finalResult = '';
-  let textBuffer = '';
-  // Track text across all conversation turns so multi-step agents (e.g., task + /simplify)
-  // preserve all summaries instead of only the final one
-  const textSections = [];
-  let currentTextSection = '';
-  // Track active tool blocks by index for input accumulation
-  const activeTools = new Map(); // index -> { name, inputJson }
-
-  // Commit accumulated text as a section (called at result events and stream end).
-  // The committed section represents an agent turn's final wrap-up.
-  const commitSection = () => {
-    const section = currentTextSection.trim();
-    if (section) {
-      textSections.push(section);
-      currentTextSection = '';
-    }
-  };
-
-  // At a tool-call boundary the accumulated text is interim narration ("Now let me…")
-  // that gets superseded by whatever the agent says after the tool returns. Discard it
-  // so only the final post-last-tool wrap-up survives into textSections.
-  const discardSection = () => { currentTextSection = ''; };
-
-  const processChunk = (rawData) => {
-    const lines = [];
-    lineBuffer += rawData;
-
-    // Split on newlines - each JSON object is on its own line
-    const parts = lineBuffer.split('\n');
-    // Keep the last incomplete line in the buffer
-    lineBuffer = parts.pop() || '';
-
-    for (const part of parts) {
-      const trimmed = part.trim();
-      if (!trimmed) continue;
-
-      // Skip non-JSON lines (stderr mixed in, etc.)
-      if (!trimmed.startsWith('{')) continue;
-      const parsed = safeParse(trimmed);
-      if (!parsed) continue;
-
-      // Extract text from streaming deltas
-      if (parsed.type === 'stream_event') {
-        const event = parsed.event;
-        if (event?.type === 'content_block_delta' && event.delta?.type === 'text_delta') {
-          const text = event.delta.text;
-          textBuffer += text;
-          currentTextSection += text;
-          // Emit complete lines for readability, accumulate partial
-          const textLines = textBuffer.split('\n');
-          textBuffer = textLines.pop() || '';
-          for (const tl of textLines) {
-            lines.push(tl);
-          }
-        }
-        // Accumulate tool input JSON deltas
-        if (event?.type === 'content_block_delta' && event.delta?.type === 'input_json_delta') {
-          const idx = event.index;
-          const tool = activeTools.get(idx);
-          if (tool) {
-            tool.inputJson += event.delta.partial_json || '';
-          }
-        }
-        // Track tool use start - record name and begin accumulating input
-        if (event?.type === 'content_block_start' && event.content_block?.type === 'tool_use') {
-          const toolName = event.content_block.name || 'unknown';
-          const idx = event.index;
-          activeTools.set(idx, { name: toolName, inputJson: '' });
-          lines.push(`🔧 Using ${toolName}...`);
-          discardSection();
-        }
-        // When tool input is complete, emit a detailed summary line
-        if (event?.type === 'content_block_stop') {
-          const idx = event.index;
-          const tool = activeTools.get(idx);
-          if (tool) {
-            if (tool.inputJson) {
-              const input = safeParse(tool.inputJson);
-              if (input) {
-                const detail = summarizeToolInput(tool.name, input);
-                if (detail) {
-                  lines.push(`  → ${detail}`);
-                }
-              }
-            }
-            activeTools.delete(idx);
-          }
-        }
-      }
-
-      // Extract tool results from assistant messages
-      if (parsed.type === 'assistant') {
-        const content = parsed.message?.content;
-        if (Array.isArray(content)) {
-          for (const block of content) {
-            if (block.type === 'tool_result' && typeof block.content === 'string') {
-              const firstLine = block.content.split('\n')[0]?.substring(0, 200);
-              if (firstLine) {
-                lines.push(`  ↳ ${firstLine}`);
-              }
-            }
-          }
-        }
-      }
-
-      // Capture final result text for output file
-      if (parsed.type === 'result') {
-        if (textBuffer) {
-          lines.push(textBuffer);
-          textBuffer = '';
-        }
-        commitSection();
-        finalResult = parsed.result || '';
-      }
-    }
-
-    return lines;
-  };
-
-  const flush = () => {
-    const lines = [];
-    if (textBuffer) {
-      lines.push(textBuffer);
-      textBuffer = '';
-    }
-    commitSection();
-    return lines;
-  };
-
-  // Multi-section: return all text turns combined (e.g., task summary + simplify summary)
-  // Single-section: return the CLI result field (cleaner, no tool call noise)
-  const getFinalResult = () => {
-    if (textSections.length > 1) {
-      return textSections.join('\n\n');
-    }
-    return finalResult;
-  };
-
-  return { processChunk, flush, getFinalResult };
-}
 
 /**
  * Build spawn command and arguments for a CLI provider.
